pyproject.toml
[project]
name = "recipes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "flask-cors>=5.0.0",
    "fuzzywuzzy>=0.18.0",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "openai>=1.109.1",
    "playwright>=1.55.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.0.1",
    "python-levenshtein>=0.26.1",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.43",
    "sqlmodel>=0.0.25",
    "tenacity>=9.1.2",
]

[dependency-groups]
dev = [
    "jupyterlab>=4.4.5",
    "pytest>=8.4.2",
]

# Frontend commands: use bun directly
# Development: `bun --cwd frontend run dev`
# Build: `bun --cwd frontend run build`
